% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bait.mr.boundaries.R
\name{bait.mr.boundaries}
\alias{bait.mr.boundaries}
\title{Prepare a file request for Marine Regions boundaries data}
\usage{
bait.mr.boundaries(file)
}
\arguments{
\item{file}{Character. The file you want to retrieve. Available files are 'eez_12nm', 'eez_24nm', and 'eez'.}
}
\value{
A character scalar (bait) to pass to `fish()`. The scalar contains the endpoint of the http GET request.
}
\description{
Preliminary function to prepare a 'bait' (API endpoint) to `fish()` for GFW fishing effort data via fishRman's API.
}
\examples{
\donttest{
bait <- bait.mr.boundaries(file = "eez")

print(bait)

# "/mr?file=eez_boundaries_v11"
}
}
\seealso{
[fish()]
}
