% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\name{fish}
\alias{fish}
\title{Send a query request for data}
\usage{
fish(bait, sql = NULL)
}
\arguments{
\item{bait}{A 'bait' character scalar (endpoint) prepared via a bait function (e.g. `bait.gfw.effort()`).}

\item{sql}{Optional. A character scalar. Defaults to NULL. Select sql="count" to get the number of entries that the bait would normally return, sql="query" to get the query that will be run on the database.}
}
\value{
The result of your query, usually a dataframe.
}
\description{
Function to 'fish' for data using a 'bait' via fishRman's API.
}
\examples{
\dontrun{
bait <- bait.gfw.effort(table = "fishing_effort_v2", min_lat = 0, flag = c("ITA", "FRA"))

what_is_the_bait_selecting <- fish(bait = bait, sql = "query")

how_many_rows_would_catch_have <- fish(bait = bait, sql = "count")

catch <- fish(bait = bait)
}

}
\seealso{
[bait.gfw.effort()]
}
