% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\name{fish}
\alias{fish}
\title{Send a query request for data}
\usage{
fish(bait, JSON = FALSE)
}
\arguments{
\item{bait}{A 'bait' packaged via a bait function (e.g. `bait.gfw.effort()`)}

\item{JSON}{Logical. Defaults to FALSE. Whether the result should be in JSON format (text).}
}
\value{
The result of your query, usually a dataframe.
}
\description{
Function to 'fish' for data using a 'bait' via fishRman's API.
}
\examples{
bait <- bait.gfw.effort(table = "fishing_effort_v2", min_lat = 0, flag = c("ITA", "FRA"))

catch <- fish(bait = bait)

}
\seealso{
[bait.gfw.effort()]
}
